+++ NEWS.md
@@ -1,6 +1,15 @@
 History of Cuba.jl
 ==================
 
+v0.6.0 (2018-??-??)
+-------------------
+
+### Breaking Changes
+
+* Support for Julia 0.6 was dropped.
+* Keyword arguments `reltol` and `abstol` are now called, respectively, `rtol`
+  and `atol` , to match keywords in `isapprox` function.
+
 v0.5.0 (2018-05-15)
 -------------------